enhance ioctl handling for FreeRTOS/newlib device drivers

+++ src/freertos_drivers/common/Can.cxx
@@ -162,7 +162,7 @@ ssize_t Can::write(File *file, const void *buf, size_t count)
  * @param key ioctl key
  * @param data key data
  */
-int Can::ioctl(File *file, Node *node, int key, unsigned long data)
+int Can::ioctl(File *file, Node *node, unsigned long int key, unsigned long data)
 {
     return 0;
 }
