+++ src/freertos_drivers/common/Serial.hxx
@@ -118,7 +118,7 @@ private:
     * @param key ioctl key
     * @param data key data
     */
-    static int ioctl(File *file, Node *node, int key, unsigned long data);
+    static int ioctl(File *file, Node *node, unsigned long int key, unsigned long data);
 
     OSMutex mutex; /**< mutual exclusion for the device */
     Devtab devtab; /**< device tabel entry for this instance */
